updated database and added android test project

+++ MedicineReminder/src/com/example/medicinereminder/sql/DatabaseAccess.java
@@ -13,8 +13,11 @@ public class DatabaseAccess {
 	private SQLiteDatabase database;
 	private MySQLiteHelper dbHelper;
 	private String[] allColumnsUser = { MySQLiteHelper.COLUMN_ID, 
-			MySQLiteHelper.COLUMN_NAME, MySQLiteHelper.COLUMN_AGE,
+			MySQLiteHelper.COLUMN_NAME, MySQLiteHelper.COLUMN_LAST_NAME,
+			MySQLiteHelper.COLUMN_AGE,
 			MySQLiteHelper.COLUMN_VIRAL, MySQLiteHelper.COLUMN_CD4,
+			MySQLiteHelper.COLUMN_DIAG_DAY, MySQLiteHelper.COLUMN_DIAG_MONTH,
+			MySQLiteHelper.COLUMN_DIAG_YEAR,
 			MySQLiteHelper.COLUMN_DOCTOR, MySQLiteHelper.COLUMN_DOCTOR_NO};
 	private String[] allColumnsMedicine = { MySQLiteHelper.COLUMN_MEDICINE_ID, 
 			MySQLiteHelper.COLUMN_MEDICINE, MySQLiteHelper.COLUMN_DOSAGE };
@@ -29,6 +32,11 @@ public class DatabaseAccess {
 	private String[] allColumnsLog = {MySQLiteHelper.COLUMN_MESSAGE_ID,
 			MySQLiteHelper.COLUMN_MESSAGE,
 			MySQLiteHelper.COLUMN_TIME};
+	private String[] allColumnsRefill = {MySQLiteHelper.COLUMN_REFILL_ID,
+			MySQLiteHelper.COLUMN_MEDICINE_ID, MySQLiteHelper.COLUMN_MINUTE,
+			MySQLiteHelper.COLUMN_HOUR, MySQLiteHelper.COLUMN_DAY,
+			MySQLiteHelper.COLUMN_MONTH, MySQLiteHelper.COLUMN_YEAR};		
+	
 	
 	
 	public DatabaseAccess(Context context){
@@ -43,27 +51,24 @@ public class DatabaseAccess {
 		dbHelper.close();
 	}
 	
-	public void printToConsole(){
-		List<CursorHolder> allData = getAllData();
-		for (CursorHolder ch : allData){
-			System.out.println(printCursor(ch));
-		}
-		
-	}
 	
-	public String printCursor(CursorHolder ch){
-		return "userID: " + Integer.toString(ch.getInt(0)) + " userName: " + ch.getString(1); 
-		
-	}
-	
-	public CursorHolder addUserData(String name, int age,
-			int viral_load, int cd4, String doctor,
+	public CursorHolder addUserData(String name, String last_name,int age,
+			int viral_load, int cd4, int diag_day, int diag_month,
+			int diag_year, String doctor,
 			String doctor_number){
+		if (name==null || last_name==null || age<0 || age>120 || viral_load<0 ||
+			cd4<0 || diag_day<1 || diag_day >31 || diag_month<1 || diag_month > 12 ||
+			diag_year<1910 ||diag_year > 2013 || doctor ==null || doctor_number==null)
+			return null;
 		ContentValues values = new ContentValues();
 		values.put(MySQLiteHelper.COLUMN_NAME, name);
+		values.put(MySQLiteHelper.COLUMN_LAST_NAME, last_name);
 		values.put(MySQLiteHelper.COLUMN_AGE, age);
 		values.put(MySQLiteHelper.COLUMN_VIRAL, viral_load);
 		values.put(MySQLiteHelper.COLUMN_CD4, cd4);
+		values.put(MySQLiteHelper.COLUMN_DIAG_DAY, diag_day);
+		values.put(MySQLiteHelper.COLUMN_DIAG_MONTH, diag_month);
+		values.put(MySQLiteHelper.COLUMN_DIAG_YEAR, diag_year);
 		values.put(MySQLiteHelper.COLUMN_DOCTOR, doctor);
 		values.put(MySQLiteHelper.COLUMN_DOCTOR_NO, doctor_number);
 		long insertId = database.insert(MySQLiteHelper.TABLE_USERINFO, null,
@@ -79,12 +84,16 @@ public class DatabaseAccess {
 	}
 	
 	public void deleteNameEntry(String name) {
+		if( name == null)
+			return;
 	    System.out.println("UserDataEntry deleted with name: " + name);
 	    database.delete(MySQLiteHelper.TABLE_USERINFO, MySQLiteHelper.COLUMN_NAME
 	        + " = '" + name + "'", null);
 	 }
 	
 	public CursorHolder getUserData(String name){
+		if(name==null)
+			return null;
 		Cursor cursor = database.query(MySQLiteHelper.TABLE_USERINFO, allColumnsUser,
 				MySQLiteHelper.COLUMN_NAME + " = '" + name +"'", null, null, null, null);
 		cursor.moveToFirst();
@@ -110,6 +119,8 @@ public class DatabaseAccess {
 	
 	
 	public CursorHolder addMedicine(String medicine, int dosage){
+		if(medicine==null|| dosage < 0 )
+			return null;
 		ContentValues values = new ContentValues();
 		values.put(MySQLiteHelper.COLUMN_MEDICINE, medicine);
 		values.put(MySQLiteHelper.COLUMN_DOSAGE, dosage);
@@ -126,12 +137,16 @@ public class DatabaseAccess {
 	}
 	
 	public void deleteMedicine(String medicine) {
+		if(medicine==null)
+			return;
 	    System.out.println("Medicine deleted with nane: " + medicine);
 	    database.delete(MySQLiteHelper.TABLE_MEDICINE, MySQLiteHelper.COLUMN_MEDICINE
 	        + " = '" + medicine + "'", null);
 	 }
 	
 	public CursorHolder getMedicine(String medicine){
+		if(medicine==null)
+			return null;
 		Cursor cursor = database.query(MySQLiteHelper.TABLE_MEDICINE, allColumnsMedicine,
 				MySQLiteHelper.COLUMN_MEDICINE + " = '" + medicine +"'", null, null, null, null);
 		cursor.moveToFirst();
@@ -157,6 +172,10 @@ public class DatabaseAccess {
 	
 	public CursorHolder addSchedule(int medicine_id, int minute,
 			int hour, int day, int month, int year){
+		if(medicine_id<1 || minute <0 || minute > 59 || hour < 1 ||
+				hour > 12 || day < 1 || day > 31 || month < 1 ||
+				month > 12 || year < 1910 || year > 2013)
+			return null;
 		ContentValues values = new ContentValues();
 		values.put(MySQLiteHelper.COLUMN_MEDICINE_ID, medicine_id);
 		values.put(MySQLiteHelper.COLUMN_MINUTE, minute);
@@ -177,12 +196,16 @@ public class DatabaseAccess {
 	}
 	
 	public void deleteSchedule(int schedule_id) {
+		if(schedule_id < 1)
+			return;
 	    System.out.println("Schedule deleted with id: " + schedule_id);
 	    database.delete(MySQLiteHelper.TABLE_SCHEDULE, MySQLiteHelper.COLUMN_SCHEDULE_ID
 	        + " = " + schedule_id, null);
 	 }
 	
 	public CursorHolder getSchedule(int schedule_id) {
+		if(schedule_id < 1)
+			return null;
 		Cursor cursor = database.query(MySQLiteHelper.TABLE_SCHEDULE, allColumnsSchedule,
 				MySQLiteHelper.COLUMN_SCHEDULE_ID + " = " + schedule_id, null, null, null, null);
 		cursor.moveToFirst();
@@ -208,6 +231,10 @@ public class DatabaseAccess {
 	
 	public CursorHolder addAppt(String clinic, int minute,
 			int hour, int day, int month, int year){
+		if(clinic==null || minute <0 || minute > 59 || hour < 1 ||
+				hour > 12 || day < 1 || day > 31 || month < 1 ||
+				month > 12 || year < 1910 || year > 2013)
+			return null;
 		ContentValues values = new ContentValues();
 		values.put(MySQLiteHelper.COLUMN_CLINIC, clinic);
 		values.put(MySQLiteHelper.COLUMN_MINUTE, minute);
@@ -228,13 +255,17 @@ public class DatabaseAccess {
 	}
 	
 	public void deleteAppt(int appt_id) {
+		if( appt_id < 1)
+			return;
 	    System.out.println("Appointment deleted with id: " + appt_id);
 	    database.delete(MySQLiteHelper.TABLE_APPT, MySQLiteHelper.COLUMN_APPT_ID
 	        + " = " + appt_id, null);
 	 }
 	
 	public CursorHolder getAppt(int appt_id) {
-		Cursor cursor = database.query(MySQLiteHelper.TABLE_SCHEDULE, allColumnsAppt,
+		if( appt_id < 1)
+			return null;
+		Cursor cursor = database.query(MySQLiteHelper.TABLE_APPT, allColumnsAppt,
 				MySQLiteHelper.COLUMN_APPT_ID + " = " + appt_id, null, null, null, null);
 		cursor.moveToFirst();
 		CursorHolder newData = cursorToAppt(cursor);
@@ -258,6 +289,8 @@ public class DatabaseAccess {
 	}
 	
 	public CursorHolder addLog(String message, String time){
+		if(message == null || time ==null)
+			return null;
 		ContentValues values = new ContentValues();
 		values.put(MySQLiteHelper.COLUMN_MESSAGE, message);
 		values.put(MySQLiteHelper.COLUMN_TIME, time);
@@ -273,10 +306,21 @@ public class DatabaseAccess {
 		return newData;	
 	}
 	
+	public CursorHolder getLog(int log_id) {
+		if(log_id<1)
+			return null;
+		Cursor cursor = database.query(MySQLiteHelper.TABLE_LOG, allColumnsLog,
+				MySQLiteHelper.COLUMN_MESSAGE_ID+ " = " + log_id, null, null, null, null);
+		cursor.moveToFirst();
+		CursorHolder newData = cursorToLog(cursor);
+		cursor.close();
+		return newData;
+	}
+	
 	public List<CursorHolder> getAllLogs(){
 		List<CursorHolder> logs = new ArrayList<CursorHolder>();	
-		Cursor cursor = database.query(MySQLiteHelper.TABLE_APPT,
-		        allColumnsAppt, null, null, null, null, null);
+		Cursor cursor = database.query(MySQLiteHelper.TABLE_LOG,
+		        allColumnsLog, null, null, null, null, null);
 		cursor.moveToFirst();
 		while (!cursor.isAfterLast()) {
 			CursorHolder log= cursorToLog(cursor);
@@ -288,17 +332,138 @@ public class DatabaseAccess {
 		return logs;
 	}
 	
+	public CursorHolder addRefill(int medicine_id, int minute,
+			int hour, int day, int month, int year){
+		if(medicine_id<1 || minute <0 || minute > 59 || hour < 1 ||
+				hour > 12 || day < 1 || day > 31 || month < 1 ||
+				month > 12 || year < 1910 || year > 2013)
+			return null;
+		ContentValues values = new ContentValues();
+		values.put(MySQLiteHelper.COLUMN_MEDICINE_ID, medicine_id);
+		values.put(MySQLiteHelper.COLUMN_MINUTE, minute);
+		values.put(MySQLiteHelper.COLUMN_HOUR, hour);
+		values.put(MySQLiteHelper.COLUMN_DAY, day);
+		values.put(MySQLiteHelper.COLUMN_MONTH, month);
+		values.put(MySQLiteHelper.COLUMN_YEAR, year);
+		long insertId = database.insert(MySQLiteHelper.TABLE_REFILL, null,
+				values);
+		int insert = ((Long)insertId).intValue();
+		Cursor cursor = database.query(MySQLiteHelper.TABLE_REFILL,
+				allColumnsRefill, MySQLiteHelper.COLUMN_REFILL_ID + " = " + insert,
+				null,null,null,null);
+		cursor.moveToFirst();
+		CursorHolder newData = cursorToSchedule(cursor);
+		cursor.close();
+		return newData;	
+	}
+	
+	public void deleteRefill(int refill_id) {
+		if(refill_id < 1)
+			return;
+	    System.out.println("Schedule deleted with id: " + refill_id);
+	    database.delete(MySQLiteHelper.TABLE_REFILL, MySQLiteHelper.COLUMN_REFILL_ID
+	        + " = " + refill_id, null);
+	 }
+	
+	public CursorHolder getRefill(int refill_id) {
+		if(refill_id < 1)
+			return null;
+		Cursor cursor = database.query(MySQLiteHelper.TABLE_REFILL, allColumnsRefill,
+				MySQLiteHelper.COLUMN_REFILL_ID + " = " + refill_id, null, null, null, null);
+		cursor.moveToFirst();
+		CursorHolder newData = cursorToSchedule(cursor);
+		cursor.close();
+		return newData;
+	}
+	
+	public List<CursorHolder> getAllRefills(){
+		List<CursorHolder> scheds = new ArrayList<CursorHolder>();	
+		Cursor cursor = database.query(MySQLiteHelper.TABLE_REFILL,
+		        allColumnsRefill, null, null, null, null, null);
+		cursor.moveToFirst();
+		while (!cursor.isAfterLast()) {
+			CursorHolder schedule= cursorToSchedule(cursor);
+		      scheds.add(schedule);
+		      cursor.moveToNext();
+		}
+		 // Make sure to close the cursor
+		cursor.close(); 
+		return scheds;
+	}
+	
+	//Print contents of the database
+	public void printDatabase(){
+		List<CursorHolder> temp_list = getAllData();
+		for (CursorHolder item : temp_list){
+			System.out.println("Id: " + item.getInt(0) + " Name: " +
+					item.getString(1) + " Last: " + item.getString(2) +
+					" Age: " + item.getInt(3) + " Viral Load: " + 
+					item.getInt(4) + " CD4 Count : " + item.getInt(5) +
+					" Diag Day " + item.getInt(6) + " Diag Month " +
+					item.getInt(7) + " Diag Year " + item.getInt(8) +
+					" Provider: " + item.getString(9) + " Provider No.: " +
+					item.getString(10));
+		}
+		temp_list = getAllMedicine();
+		for (CursorHolder item : temp_list){
+			System.out.println("Medicine Id: " + item.getInt(0) + " Medicine: " +
+					item.getString(1) + " Dosage: " + item.getInt(2));
+		}
+		temp_list = getAllSchedules();
+		for (CursorHolder item : temp_list){
+			System.out.println("Schedule Id: " + item.getInt(0) + 
+					"Medicine Id: " + item.getInt(1) + " Minute: " +
+					item.getInt(2) + " Hour: " + item.getInt(3) +
+					" Day " + item.getInt(4) + " Month: " + item.getInt(5) +
+					" Year: " + item.getInt(6));
+		}
+		temp_list = getAllAppts();
+		for (CursorHolder item : temp_list){
+			System.out.println("Appointment Id: " + item.getInt(0) + 
+					"Clinic: " + item.getString(1) + " Minute: " +
+					item.getInt(2) + " Hour: " + item.getInt(3) +
+					" Day " + item.getInt(4) + " Month: " + item.getInt(5) +
+					" Year: " + item.getInt(6));
+		}
+		temp_list = getAllRefills();
+		for (CursorHolder item : temp_list){
+			System.out.println("Refill Id: " + item.getInt(0) + 
+					"Medicine Id: " + item.getInt(1) + " Minute: " +
+					item.getInt(2) + " Hour: " + item.getInt(3) +
+					" Day " + item.getInt(4) + " Month: " + item.getInt(5) +
+					" Year: " + item.getInt(6));
+		}
+		temp_list = getAllLogs();
+		for (CursorHolder item : temp_list){
+			System.out.println("Message Id: " + item.getInt(0) + " Log: " +
+					item.getString(1) + " Time: " + item.getString(2));
+		}
+	}
+	
+	public void clearTables(){
+	    database.delete(MySQLiteHelper.TABLE_USERINFO, null,null);
+	    database.delete(MySQLiteHelper.TABLE_MEDICINE, null,null);
+	    database.delete(MySQLiteHelper.TABLE_SCHEDULE, null,null);
+	    database.delete(MySQLiteHelper.TABLE_APPT, null,null);
+	    database.delete(MySQLiteHelper.TABLE_LOG, null,null);
+	    database.delete(MySQLiteHelper.TABLE_REFILL, null,null);
+	    System.out.println("All tables cleared");
+	}
 	
 	
 	private CursorHolder cursorToData(Cursor cursor) {
 		CursorHolder data = new CursorHolder();
 	    data.add(cursor.getInt(0));
 	    data.add(cursor.getString(1));
-	    data.add(cursor.getInt(2));
+	    data.add(cursor.getString(2));
 	    data.add(cursor.getInt(3));
 	    data.add(cursor.getInt(4));
-	    data.add(cursor.getString(5));
-	    data.add(cursor.getString(6));
+	    data.add(cursor.getInt(5));
+	    data.add(cursor.getInt(6));
+	    data.add(cursor.getInt(7));
+	    data.add(cursor.getInt(8));
+	    data.add(cursor.getString(9));
+	    data.add(cursor.getString(10));
 	    return data;
 	}
